fix types on model validation and it's migrations

+++ street-goods-burger-backend/db/migrate/20220611143148_create_food_orders.rb
@@ -4,7 +4,7 @@ class CreateFoodOrders < ActiveRecord::Migration[7.0]
       t.string :food_name, null: false, default: ''
       t.string :food_category, null: false, default: ''
       t.integer :quantity, null: false, default: 1
-      t.bigint :price, null: false, default: 0.00
+      t.decimal :price, null: false, default: 0.00
 
       t.timestamps
     end
